[refactor] 온디맨드 설치 경계 분리
