+++ src/setup/on-demand.ts
@@ -1,9 +1,6 @@
-import { spawn } from "node:child_process"
-import { lstatSync, readFileSync } from "node:fs"
-import { dirname, join, resolve } from "node:path"
-import { homedir, platform } from "node:os"
+import { dirname, resolve } from "node:path"
+import { homedir } from "node:os"
 import { fileURLToPath } from "node:url"
-import { z } from "zod"
 import {
   getLawCredentialPath,
   readStoredLawApiKey,
@@ -11,53 +8,12 @@ import {
 } from "../lib/law-credential.js"
 import { promptSecret } from "./secret-prompt.js"
 import type { SetupOptions } from "./setup-options.js"
-
-const SKILLS_CLI_NAME = "skills"
-const SKILL_NAME = "korean-law"
-const SKILLS_CLI_VERSION = "1.5.18"
-const TARGET_AGENTS = ["claude-code", "codex"] as const
-const TARGET_SKILL_DIRECTORIES = [
-  [".claude", "skills", SKILL_NAME],
-  [".agents", "skills", SKILL_NAME],
-] as const
-const INSTALLER_ENV_KEYS = [
-  "PATH",
-  "Path",
-  "HOME",
-  "USERPROFILE",
-  "APPDATA",
-  "LOCALAPPDATA",
-  "XDG_CONFIG_HOME",
-  "SystemRoot",
-  "ComSpec",
-  "PATHEXT",
-  "TEMP",
-  "TMP",
-  "TMPDIR",
-  "SHELL",
-  "TERM",
-  "COLORTERM",
-  "LANG",
-  "LC_ALL",
-] as const
-
-interface CommandInvocation {
-  readonly command: string
-  readonly args: readonly string[]
-  readonly env: Readonly<NodeJS.ProcessEnv>
-}
+import { installGlobalSkill } from "./skill-installer.js"
 
 interface OnDemandRuntime {
   readonly credentialPath: string
   readonly userHome?: string
 }
-
-const SkillListSchema = z.array(z.object({
-  name: z.string(),
-  path: z.string(),
-  scope: z.literal("global"),
-  agents: z.array(z.string()),
-}))
 
 function writeLine(message = ""): void {
   process.stdout.write(`${message}\n`)
@@ -71,144 +27,6 @@ function getEnvironmentApiKey(): string {
   return process.env.LAW_OC?.trim() ||
     process.env.KOREAN_LAW_API_KEY?.trim() ||
     ""
-}
-
-export function buildSkillInstallInvocation(
-  packageRoot: string,
-  os: NodeJS.Platform = platform(),
-  sourceEnv: NodeJS.ProcessEnv = process.env,
-  nodeExecutable = process.execPath
-): CommandInvocation {
-  const env = Object.fromEntries(
-    INSTALLER_ENV_KEYS.flatMap((key) => {
-      const value = sourceEnv[key]
-      return value === undefined ? [] : [[key, value]]
-    })
-  )
-  return {
-    command: resolve(
-      dirname(nodeExecutable),
-      os === "win32" ? "npx.cmd" : "npx"
-    ),
-    args: [
-      "--yes",
-      "--ignore-scripts",
-      `${SKILLS_CLI_NAME}@${SKILLS_CLI_VERSION}`,
-      "add",
-      packageRoot,
-      "--global",
-      "--skill",
-      SKILL_NAME,
-      "--agent",
-      ...TARGET_AGENTS,
-      "--yes",
-      "--copy",
-    ],
-    env,
-  }
-}
-
-function buildSkillListInvocation(
-  installInvocation: CommandInvocation
-): CommandInvocation {
-  return {
-    command: installInvocation.command,
-    args: [
-      "--yes",
-      "--ignore-scripts",
-      `${SKILLS_CLI_NAME}@${SKILLS_CLI_VERSION}`,
-      "list",
-      "--global",
-      "--json",
-    ],
-    env: installInvocation.env,
-  }
-}
-
-async function runCommand(invocation: CommandInvocation): Promise<void> {
-  await new Promise<void>((resolvePromise, reject) => {
-    const child = spawn(invocation.command, invocation.args, {
-      stdio: "inherit",
-      env: invocation.env,
-    })
-    child.once("error", reject)
-    child.once("exit", (code) => {
-      if (code === 0) {
-        resolvePromise()
-        return
-      }
-      reject(new Error(`Skill installer exited with code ${code ?? "unknown"}`))
-    })
-  })
-}
-
-export async function captureCommand(invocation: CommandInvocation): Promise<string> {
-  return await new Promise<string>((resolvePromise, reject) => {
-    const child = spawn(invocation.command, invocation.args, {
-      stdio: ["ignore", "pipe", "pipe"],
-      env: invocation.env,
-    })
-    const stdout: Buffer[] = []
-    const stderr: Buffer[] = []
-    child.stdout.on("data", (chunk: Buffer) => stdout.push(chunk))
-    child.stderr.on("data", (chunk: Buffer) => stderr.push(chunk))
-    child.once("error", reject)
-    child.once("close", (code) => {
-      if (code === 0) {
-        resolvePromise(Buffer.concat(stdout).toString("utf8"))
-        return
-      }
-      const detail = Buffer.concat(stderr).toString("utf8").trim()
-      reject(new Error(
-        `Skill verification exited with code ${code ?? "unknown"}${detail ? `: ${detail}` : ""}`
-      ))
-    })
-  })
-}
-
-export function getExpectedGlobalSkillFiles(userHome: string): readonly string[] {
-  return TARGET_SKILL_DIRECTORIES.map((segments) =>
-    join(userHome, ...segments, "SKILL.md")
-  )
-}
-
-function isExpectedSkillFile(skillFile: string): boolean {
-  try {
-    const stat = lstatSync(skillFile)
-    const content = readFileSync(skillFile, "utf8")
-    return stat.isFile() &&
-      !stat.isSymbolicLink() &&
-      /^---\s*\nname: korean-law$/m.test(content)
-  } catch {
-    return false
-  }
-}
-
-function verifySkillFiles(userHome: string): void {
-  for (const skillFile of getExpectedGlobalSkillFiles(userHome)) {
-    if (isExpectedSkillFile(skillFile)) continue
-    throw new Error(
-      `Skill 파일 검증 실패 / Skill file verification failed: ${skillFile}`
-    )
-  }
-}
-
-async function verifySkillInstallation(
-  installInvocation: CommandInvocation,
-  userHome: string
-): Promise<void> {
-  const output = await captureCommand(buildSkillListInvocation(installInvocation))
-  const installed = SkillListSchema.parse(JSON.parse(output))
-  const listed = installed.some((skill) =>
-    skill.name === SKILL_NAME
-  )
-  if (!listed) {
-    throw new Error(
-      "전역 Skill 목록에서 korean-law를 확인하지 못했습니다. / " +
-      "korean-law was not found in the global Skill list."
-    )
-  }
-  verifySkillFiles(userHome)
 }
 
 async function configureCredential(configPath: string): Promise<void> {
@@ -246,9 +64,7 @@ export async function runOnDemandSetup(
   } else {
     writeLine("  + Claude Code·Codex 전역 Skill을 설치합니다.")
     writeLine("    Installing the global Skill for Claude Code and Codex.")
-    const invocation = buildSkillInstallInvocation(getPackageRoot())
-    await runCommand(invocation)
-    await verifySkillInstallation(invocation, runtime.userHome ?? homedir())
+    await installGlobalSkill(getPackageRoot(), runtime.userHome ?? homedir())
     writeLine("  + 두 전역 설치 경로를 검증했습니다. / Verified both global install paths.")
   }
